+++ mail/mail-common/src/models/rollback_item.rs
@@ -3,12 +3,12 @@ use crate::datatypes::RollbackItemType;
 use crate::datatypes::dependencies::MessageOrConversationDependencyFetcher;
 use crate::models::{Conversation, Message, MessageSyncDecision};
 use futures::stream::{FuturesOrdered, FuturesUnordered, StreamExt};
-use proton_action_queue::queue::Queue;
 use proton_action_queue::rebase::RebaseChangeSet;
 use proton_core_api::service::ApiServiceError;
 use proton_core_api::services::proton::ProtonCore;
 use proton_core_api::services::proton::{LabelId, ProtonIdMarker};
 use proton_core_api::session::Session;
+use proton_core_common::RebasableQueue;
 use proton_core_common::models::Label;
 use proton_mail_api::services::proton::ProtonMail;
 use proton_mail_api::services::proton::common::{ConversationId, MessageId};
@@ -87,7 +87,7 @@ impl RollbackItem {
         api: &Session,
         tx: &mut impl RunTransaction,
         batch: I,
-        queue: &Queue,
+        queue: &RebasableQueue<'_>,
     ) -> Result<(), MailContextError>
     where
         I: Into<Option<usize>> + Copy,
@@ -104,7 +104,7 @@ impl RollbackItem {
         api: &Session,
         tx: &mut impl RunTransaction,
         batch: I,
-        queue: &Queue,
+        queue: &RebasableQueue<'_>,
     ) -> Result<(), MailContextError>
     where
         I: Into<Option<usize>>,
@@ -117,7 +117,7 @@ impl RollbackItem {
         api: &Session,
         tx: &mut impl RunTransaction,
         batch: I,
-        queue: &Queue,
+        queue: &RebasableQueue<'_>,
     ) -> Result<(), MailContextError>
     where
         I: Into<Option<usize>>,
@@ -130,7 +130,7 @@ impl RollbackItem {
         api: &Session,
         tx: &mut impl RunTransaction,
         batch: I,
-        queue: &Queue,
+        queue: &RebasableQueue<'_>,
     ) -> Result<(), MailContextError>
     where
         I: Into<Option<usize>>,
@@ -188,7 +188,7 @@ impl RollbackItem {
         api: &Session,
         tx_runner: &mut T,
         batch: Option<usize>,
-        #[cfg_attr(not(feature = "action_rebase"), allow(unused_variables))] queue: &Queue,
+        queue: &RebasableQueue<'_>,
     ) -> Result<(), MailContextError> {
         let items: Vec<H::RemoteId> =
             Self::find_remote_ids_by_kind(H::item_type(), tx_runner.tether())
@@ -247,7 +247,6 @@ impl RollbackItem {
                             })?;
                     }
 
-                    #[cfg(feature = "action_rebase")]
                     if let Err(e) = queue
                         .rebase_in(
                             proton_action_queue::action::ActionGroup::default(),