+++ mail/mail-tui/src/app_model/mailbox/messages.rs
@@ -265,7 +265,7 @@ impl MessagesState {
             ConversationViewOptions::All,
             ctx.user_stash(),
             ctx.session(),
-            ctx.action_queue(),
+            ctx.rebaseable_queue().await,
         )
         .await?
         else {